backend/requirements.txt
Flask-SQLAlchemy ~= 3.1
Flask-Migrate ~= 4.1
flask-cors ~= 5.0
psycopg2 ~= 2.9
marshmallow ~= 3.26
